+++ src/lib/utils/address.ts
@@ -20,14 +20,16 @@ export const getAddressTypeText = (addressType: AddressReturnType) => {
 export const bech32AddressToHex = (addr: HumanAddr): HexAddr =>
   "0x".concat(toHex(fromBech32(addr).data)) as HexAddr;
 
+export const padHexAddress = (hexAddr: HexAddr): HexAddr =>
+  `0x${hexAddr.slice(2).padStart(40, "0")}` as HexAddr;
+
+export const unpadHexAddress = (hexAddr: HexAddr) =>
+  `0x${hexAddr.slice(2).replace(/^0+/, "")}` as HexAddr;
+
 export const hexToBech32Address = (
   prefix: string,
   hexAddr: HexAddr
 ): HumanAddr => {
-  let strip = hexAddr.replace("0x", "");
-  if (strip.length < 40) strip = strip.padStart(40, "0");
+  const strip = padHexAddress(hexAddr).slice(2);
   return toBech32(prefix, fromHex(strip)) as HumanAddr;
 };
-
-export const unpadHexAddress = (hexAddr: HexAddr): HexAddr =>
-  "0x".concat(hexAddr.slice(2).replace(/^0+/, "")) as HexAddr;